refactor: 예약 등록 컴포넌트 이동
